Separate Package style Code
Before this commit, we designed code in which we have the same Package for all department classes.
In the new code, we have changed it and placed every class in its own package. And we observe no changes. Just for the betterment of the score, it's done.
We have asked about its detailing thing if the answer comes like they want more detailing then we will code it in a similar way.
So, There are chances of modification.